# promotion/src/main/java/henry/jonathan/promotion/model/CalcPromotionRequest.java
package henry.jonathan.promotion.model;

import lombok.Data;

import java.util.List;

/**
 * @author henry.jonathan
 * @since
 */
@Data
public class CalcPromotionRequest {
  private List<CartItem> cartItems;
}